+++ src/main/java/ru/vyarus/guice/ext/core/generator/DynamicClassGenerator.java
@@ -27,7 +27,10 @@ import java.lang.reflect.Modifier;
  * {@code bind(MyType.class).to(DynamicClassGenerator.generate(MyType.class))}.
  * Another option is to use {@link com.google.inject.internal.DynamicClassProvider}: annotate type with
  * {@code @ProvidedBy(DynamicClassProvider.class)} and either rely on JIT (don't register type at all) or
- * simply register type: @{code bind(MyType.class)}</p>
+ * simply register type: @{code bind(MyType.class)}.
+ * If used with injectors hierarchy or within private modules, use "anchor" dependency to prevent bubbling up
+ * for resulted binding (see for example anchor implementation for use with {@code @ProvidedBy} annotation
+ * {@link ru.vyarus.guice.ext.core.generator.anchor.GeneratorAnchorModule}).</p>
  * <p>Don't use scope annotations directly - instead wrap them into
  * {@link ru.vyarus.guice.ext.core.generator.ScopeAnnotation}, because guice doesn't allow scope definition on
  * abstract types.</p>
@@ -36,6 +39,7 @@ import java.lang.reflect.Modifier;
  * @see com.google.inject.internal.DynamicClassProvider
  * @since 10.12.2014
  */
+@SuppressWarnings("checkstyle:ClassDataAbstractionCoupling")
 public final class DynamicClassGenerator {
 
     /**
@@ -47,7 +51,7 @@ public final class DynamicClassGenerator {
     }
 
     /**
-     * Shortcut for {@link #generate(Class, Class)} method to create default scoped classes.
+     * Shortcut for {@link #generate(Class, Class, Class)} method to create default scoped classes.
      * <p>
      * Method is thread safe.
      *
@@ -60,12 +64,8 @@ public final class DynamicClassGenerator {
     }
 
     /**
-     * Generates dynamic class, which guice may use as implementation and generate proxy above it,
-     * correctly applying aop features.
-     * <p>
-     * New class will inherit type annotations and constructor with annotations
-     * (if base class use constructor injection). Also constructor inherits all annotations, including
-     * parameters annotations.
+     * Shortcut for {@link #generate(Class, Class, Class)} method to create classes with provided scope
+     * (and without extra anchor).
      * <p>
      * Method is thread safe.
      *
@@ -74,9 +74,33 @@ public final class DynamicClassGenerator {
      * @param <T>   type
      * @return implementation class for provided type (will not generate if class already exist)
      */
+    public static <T> Class<T> generate(final Class<T> type,
+                                        final Class<? extends java.lang.annotation.Annotation> scope) {
+        return generate(type, scope, null);
+    }
+
+    /**
+     * Generates dynamic class, which guice may use as implementation and generate proxy above it,
+     * correctly applying aop features.
+     * <p>
+     * New class will inherit type annotations and constructor with annotations
+     * (if base class use constructor injection). Also constructor inherits all annotations, including
+     * parameters annotations. If anchor is provided then it will be added as last constructor parameter
+     * or (when abstract type has no constructor) new constructor added with one parameter (anchor).
+     * <p>
+     * Method is thread safe.
+     *
+     * @param type   interface or abstract class
+     * @param scope  scope annotation to apply on generated class (may be null for default prototype scope)
+     * @param anchor existing binding to depend generated class on (to prevent binding bubbling up to root injector)
+     * @param <T>    type
+     * @return implementation class for provided type (will not generate if class already exist)
+     * @see ru.vyarus.guice.ext.core.generator.anchor.GeneratorAnchorModule for more details about anchor usage
+     */
     @SuppressWarnings("unchecked")
     public static <T> Class<T> generate(final Class<T> type,
-                                                     final Class<? extends java.lang.annotation.Annotation> scope) {
+                                        final Class<? extends java.lang.annotation.Annotation> scope,
+                                        final Class<?> anchor) {
         Preconditions.checkNotNull(type, "Original type required");
         Preconditions.checkArgument(type.isInterface() || Modifier.isAbstract(type.getModifiers()),
                 "Type must be interface or abstract class, but provided type is not: %s", type.getName());
@@ -95,7 +119,7 @@ public final class DynamicClassGenerator {
                 // will work if class was already generated
                 targetClass = classLoader.loadClass(targetClassName);
             } catch (ClassNotFoundException ex) {
-                targetClass = generateClass(type, targetClassName, classLoader, scope);
+                targetClass = generateClass(type, targetClassName, classLoader, scope, anchor);
             }
             return (Class<T>) targetClass;
         }
@@ -104,14 +128,15 @@ public final class DynamicClassGenerator {
     @SuppressWarnings("unchecked")
     private static Class<?> generateClass(final Class<?> type, final String targetClassName,
                                           final ClassLoader classLoader,
-                                          final Class<? extends java.lang.annotation.Annotation> scope) {
+                                          final Class<? extends java.lang.annotation.Annotation> scope,
+                                          final Class<?> anchor) {
         try {
             // have to use custom pool because original type classloader could be thrown away
             // and all cached CtClass objects would be stale
             final ClassPool classPool = new ClassPool();
             classPool.appendClassPath(new LoaderClassPath(classLoader));
 
-            final CtClass impl = generateCtClass(classPool, targetClassName, type, scope);
+            final CtClass impl = generateCtClass(classPool, targetClassName, type, scope, anchor);
             return impl.toClass(classLoader, null);
         } catch (Exception ex) {
             throw new DynamicClassException("Failed to generate class for " + type.getName(), ex);
@@ -119,10 +144,12 @@ public final class DynamicClassGenerator {
     }
 
     private static CtClass generateCtClass(final ClassPool classPool, final String targetClassName, final Class type,
-                                           final Class<? extends java.lang.annotation.Annotation> scope)
+                                           final Class<? extends java.lang.annotation.Annotation> scope,
+                                           final Class<?> anchor)
             throws Exception {
 
         final CtClass ctType = classPool.get(type.getName());
+        final CtClass ctAnchor = anchor == null ? null : classPool.getCtClass(anchor.getName());
         final CtClass impl;
         if (type.isInterface()) {
             impl = classPool.makeClass(targetClassName);
@@ -131,8 +158,12 @@ public final class DynamicClassGenerator {
             impl = classPool.makeClass(targetClassName, ctType);
             final Constructor diConstructor = findDIConstructor(type);
             if (diConstructor != null) {
-                copyConstructor(impl, ctType, diConstructor);
+                copyConstructor(impl, ctType, diConstructor, ctAnchor);
             }
+        }
+        if (anchor != null && impl.getConstructors().length == 0) {
+            // create new constructor with anchor dependency
+            createAnchorConstructor(impl, ctAnchor);
         }
         final ConstPool constPool = impl.getClassFile().getConstPool();
         final AnnotationsAttribute annotations = copyAnnotations(classPool, constPool, type);
@@ -141,22 +172,43 @@ public final class DynamicClassGenerator {
         return impl;
     }
 
-    private static void copyConstructor(final CtClass impl, final CtClass ctType, final Constructor ctor)
-            throws Exception {
+    private static void copyConstructor(final CtClass impl, final CtClass ctType, final Constructor ctor,
+                                        final CtClass anchor) throws Exception {
         final ClassPool classPool = impl.getClassPool();
         final CtClass[] parameters = JavassistUtils.convertTypes(classPool, ctor.getParameterTypes());
         final CtConstructor ctConstructor = CtNewConstructor.make(
                 parameters,
                 JavassistUtils.convertTypes(classPool, ctor.getExceptionTypes()),
                 CtNewConstructor.PASS_PARAMS, null, null, impl);
+        if (anchor != null) {
+            ctConstructor.addParameter(anchor);
+        }
         final ConstPool constPool = impl.getClassFile().getConstPool();
         final MethodInfo methodInfo = ctConstructor.getMethodInfo();
         methodInfo.addAttribute(copyAnnotations(classPool, constPool, ctor));
-        methodInfo.addAttribute(copyConstructorParametersAnnotations(classPool, constPool, ctor));
+        methodInfo.addAttribute(copyConstructorParametersAnnotations(classPool, constPool, ctor, anchor != null));
         final SignatureAttribute info = copyConstructorGenericsSignature(constPool, parameters, ctType);
         if (info != null) {
             methodInfo.addAttribute(info);
         }
+        impl.addConstructor(ctConstructor);
+    }
+
+    private static void createAnchorConstructor(final CtClass impl, final CtClass anchor)
+            throws Exception {
+        final ClassPool classPool = impl.getClassPool();
+        final CtConstructor ctConstructor = CtNewConstructor.make(
+                new CtClass[]{anchor},
+                null,
+                CtNewConstructor.PASS_NONE, null, null, impl);
+        final ConstPool constPool = impl.getClassFile().getConstPool();
+        final MethodInfo methodInfo = ctConstructor.getMethodInfo();
+        // add injection annotation
+        final AnnotationsAttribute attr = new AnnotationsAttribute(constPool, AnnotationsAttribute.visibleTag);
+        final Annotation annotation = new Annotation(constPool,
+                classPool.get(javax.inject.Inject.class.getName()));
+        attr.addAnnotation(annotation);
+        methodInfo.addAttribute(attr);
         impl.addConstructor(ctConstructor);
     }
 
@@ -173,15 +225,19 @@ public final class DynamicClassGenerator {
 
     @SuppressWarnings("PMD.AvoidInstantiatingObjectsInLoops")
     private static ParameterAnnotationsAttribute copyConstructorParametersAnnotations(
-            final ClassPool classPool, final ConstPool constPool, final Constructor ctor) throws Exception {
+            final ClassPool classPool, final ConstPool constPool, final Constructor ctor,
+            final boolean anchorAdded) throws Exception {
         final int count = ctor.getParameterTypes().length;
-        final Annotation[][] paramAnnotations = new Annotation[count][];
+        final Annotation[][] paramAnnotations = new Annotation[count + (anchorAdded ? 1 : 0)][];
         for (int i = 0; i < count; i++) {
             final java.lang.annotation.Annotation[] anns = ctor.getParameterAnnotations()[i];
             paramAnnotations[i] = new Annotation[anns.length];
             for (int j = 0; j < anns.length; j++) {
                 paramAnnotations[i][j] = JavassistUtils.copyAnnotation(classPool, constPool, anns[j]);
             }
+        }
+        if (anchorAdded) {
+            paramAnnotations[count] = new Annotation[0];
         }
         final ParameterAnnotationsAttribute paramAnns = new ParameterAnnotationsAttribute(
                 constPool, ParameterAnnotationsAttribute.visibleTag);