+++ src/test/java/ru/vyarus/guice/ext/generator/support/anchor/DynamicService.java
@@ -1,0 +1,16 @@
+package ru.vyarus.guice.ext.generator.support.anchor;
+
+import javax.inject.Inject;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 22.09.2016
+ */
+public class DynamicService {
+    @Inject
+    TestIface service;
+
+    public String hello() {
+        return service.hello();
+    }
+}