+++ src/test/java/ru/vyarus/guice/ext/generator/support/anchor/PureAbstractClass.java
@@ -1,0 +1,16 @@
+package ru.vyarus.guice.ext.generator.support.anchor;
+
+import com.google.inject.ProvidedBy;
+import com.google.inject.internal.DynamicSingletonProvider;
+import ru.vyarus.guice.ext.generator.support.aop.CustomAop;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 21.09.2016
+ */
+@ProvidedBy(DynamicSingletonProvider.class)
+public abstract class PureAbstractClass {
+
+    @CustomAop
+    public abstract String hello();
+}